added NoteListButton.js and moved the Show Notes button with corresponding event listeners, added CriminalListButton.js and moved the Show All Criminals button with corresponding event listeners, moved event listeners out of notelist to corresponding component module.
