+++ scripts/notes/NoteDialog.js
@@ -1,6 +1,9 @@
+const eventHub = document.querySelector(".container");
+const notesContainerContentTargetElement = document.querySelector(".notesContainer");
+
 // Renders the dialog elements to the DOM for each specific criminal when being called in a for/of loop.
 export const EditNoteDialogElement = (noteObject) => {
-    let contentTargetElement = document.querySelector(`#note--${noteObject.id}`);
+    const contentTargetElement = document.querySelector(`#note--${noteObject.id}`);
     contentTargetElement.innerHTML += `
         <dialog class="dialog--note" id="details--${noteObject.id}">
             <label for="note--date--edit--${noteObject.id}" class="note--date--editLabel">Date:</label>
@@ -17,3 +20,55 @@ export const EditNoteDialogElement = (noteObject) => {
 
         `
 }
+// Listens for the custom event, editDialogButtonDetailEvent, to open a corresponding dialog box.
+eventHub.addEventListener("editDialogButtonDetailEvent", event => {
+    const dialogSiblingSelector = `#editNote--${event.detail.note}+dialog`;
+    const editDialog = document.querySelector(dialogSiblingSelector);
+    editDialog.showModal();
+    const setDate = (noteObjectDate) => {
+        const dialogDateNode = editDialog.getElementsByClassName("note--date--edit");
+        dialogDateNode[0].value = `${noteObjectDate}`;
+      }
+    setDate(event.detail.date);
+})
+
+// Listens for a "click" event and invokes the function, editNote, to replace the JSON data object with new values.
+notesContainerContentTargetElement.addEventListener(
+    "click", 
+    theEditEvent => {
+        if (event.target.id.startsWith("editNoteSubmit--")) {
+            const [prefix, editedNoteId, editedCriminalNoteId] = theEditEvent.target.id.split('--');
+            const contentTargetDate = document.getElementById(`note--date--edit--${editedNoteId}`).value;
+            const contentTargetSuspect = document.getElementById(`note--suspect--edit--${editedNoteId}`).value;
+            const contentTargetNoteText = document.getElementById(`note--text--edit--${editedNoteId}`).value;
+
+            const editedNoteObject = {
+                "date": contentTargetDate,
+                "suspect": contentTargetSuspect,
+                "noteText": contentTargetNoteText,
+                "criminalId": editedCriminalNoteId,
+                "id": editedNoteId
+            }
+
+            const theEditNoteEvent = new CustomEvent("editNoteEvent", {
+                detail: {
+                    note: editedNoteObject,
+                }
+            })
+            eventHub.dispatchEvent(theEditNoteEvent);
+            document.getElementById(`details--${editedNoteId}`).close()
+        }
+    }
+)
+
+// Listens for a "click" event and closes a corresponding dialog box.
+notesContainerContentTargetElement.addEventListener(
+    "click", 
+    event => {
+    if (event.target.id.startsWith("close-")) {
+        const [prefix, chosenDialog] = event.target.id.split("-");
+        const theDialogBoxID = `#details--${chosenDialog}`;
+        const theDialogElement = document.querySelector(theDialogBoxID);
+        theDialogElement.close();
+    }
+})
